+++ src/aihwkit/simulator/rpu_base_src/rpu_base.h
@@ -12,11 +12,11 @@
 #pragma once
 #include "rpu.h"
 #include "rpu_constantstep_device.h"
-#include "rpu_difference_device.h"
 #include "rpu_expstep_device.h"
 #include "rpu_linearstep_device.h"
 #include "rpu_mixedprec_device.h"
 #include "rpu_mixedprec_device_base.h"
+#include "rpu_onesided_device.h"
 #include "rpu_powstep_device.h"
 #include "rpu_pulsed.h"
 #include "rpu_simple_device.h"