+++ CHANGELOG.md
@@ -16,10 +16,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ### Added
 
+* Weight refresh mechanism for ``OneSidedUnitCell`` to counteract
+  saturation, by differential read, reset, and re-write. (\#209)
+  
 * A number of new config presets added to the library, namely `EcRamMOPreset`,
   `EcRamMO2Preset`, `EcRamMO4Preset`, `TikiTakaEcRamMOPreset`,
   `MixedPrecisionEcRamMOPreset`. These can be used for tile configuration
   (`rpu_config`). They specify a particular device and optimizer choice. (\#207)
+
+### Changed
+* Renamed the `DifferenceUnitCell` to `OneSidedUnitCell` which more
+  properly reflects its function. (\#209)
 
 ## [0.3.0] - 2021/04/14